Rename MatrixMultiply file to MatrixMult
