+++ examples/plot_matrixmult.py
@@ -11,7 +11,7 @@ import numpy as np
 from mpi4py import MPI
 
 from pylops_mpi import DistributedArray, Partition
-from pylops_mpi.basicoperators.MatrixMultiply import MPISUMMAMatrixMult
+from pylops_mpi.basicoperators.MatrixMult import MPISUMMAMatrixMult
 
 np.random.seed(42)
